requirements.txt
langchain==0.1.9
langchain-openai==0.0.5
openai==1.12.0
python-dotenv==1.0.1
PyPDF2==3.0.1
markdown2==2.4.12
streamlit==1.31.1
google-search-results==2.4.2
requests==2.31.0
